+++ threedis/src/main/java/com/three/core/msg/mtype/perType/GiftType.java
@@ -4,5 +4,5 @@ import com.three.core.msg.mtype.NetMessageType;
 
 public final class GiftType extends NetMessageType{
 
-	public static final short CG_SEND_GIFT = ++GIFT_BEGIN;
+	public static final int CG_SEND_GIFT = ++GIFT_BEGIN;
 }
